backend/requirements.txt
# Web 框架
Flask==2.3.3
Flask-CORS==4.0.0

# AI 核心依赖
langchain==0.0.340
langchain-openai==0.0.2
pymilvus==2.2.14
torch>=2.1.0
torchvision>=0.15.0
transformers>=4.56.0
openai-clip  # 使用openai-clip替代clip

# OmniVision 多模态模型支持
# transformers>=4.56.0  # 已包含在上面
huggingface-hub>=0.34.0
tokenizers>=0.22.0
safetensors>=0.4.3

# 图像处理
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3

# 工具库
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
pandas==2.0.3

# 文件处理
werkzeug==2.3.7

# 生产服务器
gunicorn==21.2.0  # Linux/macOS
waitress==2.1.2   # Windows
